test: split web fixtures into multiple files
