+++ packages/webpack-config-web/test/core.test.js
@@ -8,15 +8,7 @@ const {readFileSync} = require("fs");
 const {join} = require("path");
 const {chromium} = require("playwright");
 const {build, getFiles} = require("../../functions");
-const {
-	getSnapshot,
-	getSnapshotLinkScriptTags,
-	getSnapshotMetaTags,
-	getSnapshotMultiple,
-	getSnapshotColor,
-	getSnapshotMultipleColor,
-	getSnapshotImage
-} = require("./functions");
+const {getSnapshot, getSnapshotLinkScriptTags, getSnapshotMetaTags, getSnapshotMultiple} = require("./functions");
 
 it("Basic", async function () {
 	const id = "basic";
@@ -493,417 +485,6 @@ it("Skip Reset: True", async function () {
 	}
 });
 
-it("CSS Modules", async function () {
-	const id = "css_modules";
-	const dist = join(process.cwd(), `test/${id}/dist`);
-	await build({id, expectFiles: ["index.html", "app-css-modules.js", "app-css-modules.css"]});
-
-	const port = 3010;
-	const app = express();
-	app.use(express.static(dist));
-	const server = app.listen(port);
-	try {
-		const actual = await getSnapshotColor(chromium, `http://localhost:${port}/`);
-		const expected = "rgb(0, 128, 0)";
-		strictEqual(actual, expected);
-	} finally {
-		server.close();
-	}
-});
-
-it("CSS Modules: Custom Filename", async function () {
-	const id = "css_modules_filename";
-	const dist = join(process.cwd(), `test/${id}/dist`);
-	await build({
-		id,
-		expectFiles: ["index.html", "app-css-modules-filename.js", "subfolder/custom.app-css-modules-filename.css"]
-	});
-
-	const port = 3011;
-	const app = express();
-	app.use(express.static(dist));
-	const server = app.listen(port);
-	try {
-		const actual = await getSnapshotColor(chromium, `http://localhost:${port}/`);
-		const expected = "rgb(0, 128, 0)";
-		strictEqual(actual, expected);
-	} finally {
-		server.close();
-	}
-});
-
-it("CSS without CSS Modules", async function () {
-	const id = "css_no_modules";
-	const dist = join(process.cwd(), `test/${id}/dist`);
-	await build({id, expectFiles: ["index.html", "app-css-no-modules.js", "app-css-no-modules.css"]});
-
-	const port = 3012;
-	const app = express();
-	app.use(express.static(dist));
-	const server = app.listen(port);
-	try {
-		const actual = await getSnapshotColor(chromium, `http://localhost:${port}/`);
-		const expected = "rgb(0, 128, 0)";
-		strictEqual(actual, expected);
-	} finally {
-		server.close();
-	}
-});
-
-it("CSS Reset", async function () {
-	const id = "css_reset";
-	const dist = join(process.cwd(), `test/${id}/dist`);
-	await build({id, expectFiles: ["index.html", "app-css-reset.js", "app-css-reset.css"]});
-
-	const port = 3013;
-	const app = express();
-	app.use(express.static(dist));
-	const server = app.listen(port);
-	try {
-		const actual = await getSnapshotColor(chromium, `http://localhost:${port}/`);
-		const expected = "rgb(0, 128, 0)";
-		strictEqual(actual, expected);
-	} finally {
-		server.close();
-	}
-});
-
-it("SCSS Reset", async function () {
-	const id = "scss_reset";
-	const dist = join(process.cwd(), `test/${id}/dist`);
-	await build({id, expectFiles: ["index.html", "app-scss-reset.js", "app-scss-reset.css"]});
-
-	const port = 3014;
-	const app = express();
-	app.use(express.static(dist));
-	const server = app.listen(port);
-	try {
-		const actual = await getSnapshotColor(chromium, `http://localhost:${port}/`);
-		const expected = "rgb(0, 128, 0)";
-		strictEqual(actual, expected);
-	} finally {
-		server.close();
-	}
-});
-
-it("SCSS Global Import", async function () {
-	const id = "scss_global_import";
-	const dist = join(process.cwd(), `test/${id}/dist`);
-	await build({id, expectFiles: ["index.html", "app-scss-global-import.js", "app-scss-global-import.css"]});
-
-	const port = 3015;
-	const app = express();
-	app.use(express.static(dist));
-	const server = app.listen(port);
-	try {
-		const actual = await getSnapshotColor(chromium, `http://localhost:${port}/`);
-		const expected = "rgb(0, 128, 0)";
-		strictEqual(actual, expected);
-	} finally {
-		server.close();
-	}
-});
-
-it("SCSS Global Define", async function () {
-	const id = "scss_global_define";
-	const dist = join(process.cwd(), `test/${id}/dist`);
-	await build({id, expectFiles: ["index.html", "app-scss-global-define.js", "app-scss-global-define.css"]});
-
-	const port = 3016;
-	const app = express();
-	app.use(express.static(dist));
-	const server = app.listen(port);
-	try {
-		const actual = await getSnapshotColor(chromium, `http://localhost:${port}/`);
-		const expected = "rgb(128, 128, 0)";
-		strictEqual(actual, expected);
-	} finally {
-		server.close();
-	}
-});
-
-it("SCSS Basic", async function () {
-	const id = "scss_basic";
-	const dist = join(process.cwd(), `test/${id}/dist`);
-	await build({id, expectFiles: ["index.html", "app-scss-basic.js", "app-scss-basic.css"]});
-
-	const port = 3017;
-	const app = express();
-	app.use(express.static(dist));
-	const server = app.listen(port);
-	try {
-		const actual = await getSnapshotColor(chromium, `http://localhost:${port}/`);
-		const expected = "rgb(0, 128, 255)";
-		strictEqual(actual, expected);
-	} finally {
-		server.close();
-	}
-});
-
-it("SCSS Import File", async function () {
-	const id = "scss_import_file";
-	const dist = join(process.cwd(), `test/${id}/dist`);
-	await build({id, expectFiles: ["index.html", "app-scss-import-file.js", "app-scss-import-file.css"]});
-
-	const port = 3018;
-	const app = express();
-	app.use(express.static(dist));
-	const server = app.listen(port);
-	try {
-		const actual = await getSnapshotColor(chromium, `http://localhost:${port}/`);
-		const expected = "rgb(0, 0, 255)";
-		strictEqual(actual, expected);
-	} finally {
-		server.close();
-	}
-});
-
-it("SCSS Import Module", async function () {
-	const id = "scss_import_module";
-	const dist = join(process.cwd(), `test/${id}/dist`);
-	await build({id, expectFiles: ["index.html", "app-scss-import-module.js", "app-scss-import-module.css"]});
-
-	const port = 3019;
-	const app = express();
-	app.use(express.static(dist));
-	const server = app.listen(port);
-	try {
-		const actual = await getSnapshotColor(chromium, `http://localhost:${port}/`);
-		const expected = "rgb(0, 255, 0)";
-		strictEqual(actual, expected);
-	} finally {
-		server.close();
-	}
-});
-
-it("CSS Chunks", async function () {
-	const id = "css_chunks";
-	const dist = join(process.cwd(), `test/${id}/dist`);
-	await build({
-		id,
-		expectFiles: ["index.html", "app-css-chunks.js", "app-css-chunks.css", "chunk.0.js", "chunk.0.css"]
-	});
-
-	const port = 3020;
-	const app = express();
-	app.use(express.static(dist));
-	const server = app.listen(port);
-	try {
-		const actual = await getSnapshotColor(chromium, `http://localhost:${port}/`);
-		const expected = "rgb(0, 0, 255)";
-		strictEqual(actual, expected);
-	} finally {
-		server.close();
-	}
-});
-
-it("CSS Chunks Filename", async function () {
-	const id = "css_chunks_filename";
-	const dist = join(process.cwd(), `test/${id}/dist`);
-	await build({
-		id,
-		expectFiles: [
-			"index.html",
-			"app-css-chunks-filename.js",
-			"app-css-chunks-filename.css",
-			"chunk.0.js",
-			"subfolder/custom.chunk.0.css"
-		]
-	});
-
-	const port = 3021;
-	const app = express();
-	app.use(express.static(dist));
-	const server = app.listen(port);
-	try {
-		const actual = await getSnapshotColor(chromium, `http://localhost:${port}/`);
-		const expected = "rgb(0, 0, 255)";
-		strictEqual(actual, expected);
-	} finally {
-		server.close();
-	}
-});
-
-it("SCSS Chunks", async function () {
-	const id = "scss_chunks";
-	const dist = join(process.cwd(), `test/${id}/dist`);
-	await build({
-		id,
-		expectFiles: ["index.html", "app-scss-chunks.js", "app-scss-chunks.css", "chunk.0.js", "chunk.0.css"]
-	});
-
-	const port = 3022;
-	const app = express();
-	app.use(express.static(dist));
-	const server = app.listen(port);
-	try {
-		const actual = await getSnapshotColor(chromium, `http://localhost:${port}/`);
-		const expected = "rgb(0, 0, 255)";
-		strictEqual(actual, expected);
-	} finally {
-		server.close();
-	}
-});
-
-it("SCSS Chunks: Custom Filename", async function () {
-	const id = "scss_chunks_filename";
-	const dist = join(process.cwd(), `test/${id}/dist`);
-	await build({
-		id,
-		expectFiles: [
-			"index.html",
-			"app-scss-chunks-filename.js",
-			"app-scss-chunks-filename.css",
-			"chunk.0.js",
-			"subfolder/custom.chunk.0.css"
-		]
-	});
-
-	const port = 3023;
-	const app = express();
-	app.use(express.static(dist));
-	const server = app.listen(port);
-	try {
-		const actual = await getSnapshotColor(chromium, `http://localhost:${port}/`);
-		strictEqual(actual, "rgb(0, 0, 255)");
-	} finally {
-		server.close();
-	}
-});
-
-it("SCSS Chunks & Variables", async function () {
-	const id = "scss_chunks_variables";
-	const dist = join(process.cwd(), `test/${id}/dist`);
-	await build({
-		id,
-		expectFiles: [
-			"index.html",
-			"app-scss-chunks-variables.js",
-			"app-scss-chunks-variables.css",
-			"chunk.0.js",
-			"chunk.0.css"
-		]
-	});
-
-	const port = 3024;
-	const app = express();
-	app.use(express.static(dist));
-	const server = app.listen(port);
-	try {
-		const actual = await getSnapshotColor(chromium, `http://localhost:${port}/`);
-		const expected = "rgb(0, 0, 128)";
-		strictEqual(actual, expected);
-	} finally {
-		server.close();
-	}
-});
-
-it("SCSS Data", async function () {
-	const id = "scss_data";
-	const dist = join(process.cwd(), `test/${id}/dist`);
-	await build({
-		id,
-		expectFiles: [
-			"index.html",
-			"app-scss-data.js"
-
-			// Data alone isn't enough to produce a .css: at least one entry must contain
-			// an import or polyfill to a stylesheet, otherwise the loader skips
-			// extracting styles, unfortunately.
-			// "app-scss-data.css"
-		]
-	});
-
-	const port = 3025;
-	const app = express();
-	app.use(express.static(dist));
-	const server = app.listen(port);
-	try {
-		const actual = await getSnapshotColor(chromium, `http://localhost:${port}/`);
-		strictEqual(actual !== "rgb(0, 255, 0)", true, "Inline style");
-	} finally {
-		server.close();
-	}
-});
-
-it("SCSS Data & Import", async function () {
-	const id = "scss_data_import";
-	const dist = join(process.cwd(), `test/${id}/dist`);
-	await build({id, expectFiles: ["index.html", "app-scss-data-import.js", "app-scss-data-import.css"]});
-
-	const port = 3026;
-	const app = express();
-	app.use(express.static(dist));
-	const server = app.listen(port);
-	try {
-		const actual = await getSnapshotMultipleColor(chromium, `http://localhost:${port}/`);
-		const expected = {
-			body: "rgb(255, 0, 0)",
-			mocha: "rgb(0, 0, 255)"
-		};
-		deepStrictEqual(actual, expected);
-	} finally {
-		server.close();
-	}
-});
-
-it("SCSS Data Function", async function () {
-	const id = "scss_data_function";
-	const dist = join(process.cwd(), `test/${id}/dist`);
-	await build({id, expectFiles: ["index.html", "app-scss-data-function.js", "app-scss-data-function.css"]});
-
-	const port = 3027;
-	const app = express();
-	app.use(express.static(dist));
-	const server = app.listen(port);
-	try {
-		const actual = await getSnapshotMultipleColor(chromium, `http://localhost:${port}/`);
-		const expected = {
-			body: "rgb(255, 0, 0)",
-			mocha: "rgb(0, 0, 255)"
-		};
-		deepStrictEqual(actual, expected);
-	} finally {
-		server.close();
-	}
-});
-
-it("SCSS & Image", async function () {
-	const id = "scss_image";
-	const dist = join(process.cwd(), `test/${id}/dist`);
-	await build({id, expectFiles: ["index.html", "app-scss-image.js", "app-scss-image.css", "assets/large.jpg"]});
-
-	const port = 3028;
-	const app = express();
-	app.use(express.static(dist));
-	const server = app.listen(port);
-	try {
-		const actual = await getSnapshotImage(chromium, `http://localhost:${port}/`);
-		const expected = `url("http://localhost:${port}/assets/large.jpg")`;
-		strictEqual(actual, expected);
-	} finally {
-		server.close();
-	}
-});
-
-it("CSS & Image", async function () {
-	const id = "css_image";
-	const dist = join(process.cwd(), `test/${id}/dist`);
-	await build({id, expectFiles: ["index.html", "app-css-image.js", "app-css-image.css", "assets/large.jpg"]});
-
-	const port = 3029;
-	const app = express();
-	app.use(express.static(dist));
-	const server = app.listen(port);
-	try {
-		const actual = await getSnapshotImage(chromium, `http://localhost:${port}/`);
-		const expected = `url("http://localhost:${port}/assets/large.jpg")`;
-		strictEqual(actual, expected);
-	} finally {
-		server.close();
-	}
-});
-
 it("Sourcemaps", async function () {
 	const id = "sourcemaps";
 	await build({
@@ -1098,109 +679,6 @@ it("Chunks: Polyfills", async function () {
 it("Skip Postprocessing", async function () {
 	const id = "skip_processing";
 	await build({id, expectFiles: ["app-skip-postprocessing.js"]});
-});
-
-it("Webworkers: Basic", async function () {
-	const id = "webworker";
-	const dist = join(process.cwd(), `test/${id}/dist`);
-	await build({id, expectFiles: ["index.html", "app-webworker.js", "myworker.webworker.js"]});
-
-	const port = 3035;
-	const app = express();
-	app.use(express.static(dist));
-	const server = app.listen(port);
-	try {
-		const actual = await getSnapshot(chromium, `http://localhost:${port}/`);
-		const expected = [
-			{
-				nodeName: "#text",
-				nodeValue: "WORKER replies HELLO"
-			}
-		];
-		deepStrictEqual(actual, expected);
-	} finally {
-		server.close();
-	}
-});
-
-it("Webworkers: Custom filename", async function () {
-	const id = "webworker_filename";
-	const dist = join(process.cwd(), `test/${id}/dist`);
-	await build({
-		id,
-		expectFiles: ["index.html", "app-webworker-filename.js", "subfolder/custom.myworker.webworker.js"]
-	});
-
-	const port = 3036;
-	const app = express();
-	app.use(express.static(dist));
-	const server = app.listen(port);
-	try {
-		const actual = await getSnapshot(chromium, `http://localhost:${port}/`);
-		const expected = [
-			{
-				nodeName: "#text",
-				nodeValue: "WORKER replies HELLO"
-			}
-		];
-		deepStrictEqual(actual, expected);
-	} finally {
-		server.close();
-	}
-});
-
-it("Webworkers: Polyfills", async function () {
-	const id = "webworker_polyfills";
-	const dist = join(process.cwd(), `test/${id}/dist`);
-	await build({id, expectFiles: ["index.html", "app-webworker-polyfills.js", "application.webworker.js"]});
-
-	const port = 3037;
-	const app = express();
-	app.use(express.static(dist));
-	const server = app.listen(port);
-	try {
-		const actual = await getSnapshotMultiple(chromium, `http://localhost:${port}/`);
-		const expected = {
-			mocha1: [
-				{
-					nodeName: "#text",
-					nodeValue: "BOTH once undefined MAIN once undefined"
-				}
-			],
-			mocha2: [
-				{
-					nodeName: "#text",
-					nodeValue: "BOTH once WORKER once undefined MODULE once"
-				}
-			]
-		};
-		deepStrictEqual(actual, expected);
-	} finally {
-		server.close();
-	}
-});
-
-it("Webworkers: No export {}", async function () {
-	const id = "webworker";
-	const dist = join(process.cwd(), `test/${id}/dist`);
-	await build({id, expectFiles: ["index.html", "app-webworker.js", "myworker.webworker.js"]});
-
-	const port = 3038;
-	const app = express();
-	app.use(express.static(dist));
-	const server = app.listen(port);
-	try {
-		const actual = await getSnapshot(chromium, `http://localhost:${port}/`);
-		const expected = [
-			{
-				nodeName: "#text",
-				nodeValue: "WORKER replies HELLO"
-			}
-		];
-		deepStrictEqual(actual, expected);
-	} finally {
-		server.close();
-	}
 });
 
 it("Fails: fs", async function () {
